add swagger ui proxy
