+++ function/template/go/Dockerfile
