+++ function/template/go/main.go
