+++ function/template/go/template.yml
