+++ go.mod
@@ -1,11 +1,10 @@
 module github.com/optum/faas-swagger
 
-go 1.13
+go 1.14
 
 require (
 	github.com/ghodss/yaml v1.0.0
-	github.com/google/go-cmp v0.3.1
-	github.com/openfaas/faas-provider v0.0.0-20191030180248-7677057e416b
-	github.com/rakyll/statik v0.1.6
-	gopkg.in/yaml.v2 v2.2.7 // indirect
+	github.com/openfaas/faas-provider v0.15.1
+	github.com/pkg/errors v0.9.1
+	github.com/rakyll/statik v0.1.7
 )
