--- k8.yaml
@@ -1,83 +1,0 @@
-apiVersion: v1
-kind: Service
-metadata:
-  name: faas-swagger
-spec:
-  type : ClusterIP
-  ports:
-  - name: http
-    port: 8080
-    targetPort: 8080
-  selector:
-    app: faas-swagger
----
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  labels:
-    app: faas-swagger
-  name: faas-swagger
-spec:
-  replicas: 1
-  selector:
-    matchLabels:
-      app: faas-swagger
-  template:
-    metadata:
-      labels:
-        app: faas-swagger
-    spec:
-      containers:
-      - name: faas-swagger
-        resources:
-          limits:
-            cpu: "100m"
-            memory: "256Mi"
-          requests:
-            cpu: "60m"
-            memory: "128Mi"
-        env:
-        - name: openfaas_gateway
-          value: "http://gateway:8080"
-        - name: swagger_file_path
-          value: "/opt/files"
-        image: <changereponame>/faas-swagger:latest
-        imagePullPolicy: Always
-        ports:
-        - containerPort: 8080
-        volumeMounts:
-        - mountPath: /opt/files
-          name: swagger-config
-      volumes:
-      - name: swagger-config
-        configMap:
-          defaultMode: 420
-          name: swagger
-        # securityContext:
-        #     runAsUser: 1001
----
-apiVersion: v1
-kind: ConfigMap
-metadata:
-  name: swagger
-data:
-  swagger.yaml: |
-    info:
-      license:
-        name: MIT
-      title: Openfaas Swagger
-      version: 1.0.0
-    openapi: 3.0.0
-    paths:
-    servers:
-    - url: /
-  sample.yaml: |
-    post:
-      requestBody:
-        content:
-          application/json:
-            schema:
-              type: object
-      responses:
-        "200":
-          description: success
