+++ static/swaggerui/index.html
@@ -39,7 +39,7 @@
     window.onload = function() {
       // Begin Swagger UI call region
       const ui = SwaggerUIBundle({
-        url: "/swagger.yaml",
+        url: "/faas-swagger-fn",
         dom_id: '#swagger-ui',
         deepLinking: true,
         presets: [
