--- swagger-as-function/faas-swagger.yml
@@ -1,13 +1,0 @@
-version: 1.0
-provider:
-  name: openfaas
-  gateway: http://localhost:8081
-functions:
-  faas-swagger-fn:
-    lang: go
-    handler: ./faas-swagger
-    image: murugappans/faas-swagger-fn:latest
-    environment:
-      openfaas_gateway: "http://gateway.openfaas:8080"
-    secrets:
-    - swagger
