--- test-files/files/out-files/test1.yaml
@@ -1,93 +1,0 @@
-info:
-  license:
-    name: MIT
-  title: Openfaas Sandbox
-  version: 1.0.0
-openapi: 3.0.0
-paths:
-  /serverdecommission:
-    post:
-      requestBody:
-        content:
-          text/plain:
-            schema:
-              properties:
-                serverName:
-                  type: string
-                userMSID:
-                  type: string
-              type: object
-        description: user to add to the system
-      responses:
-        5XX:
-          description: Internal API Error.
-        "200":
-          content:
-            text/plain:
-              schema:
-                example: Server Decommission successful
-                type: string
-          description: Decommission message
-        "404":
-          description: Invalid server.
-        "417":
-          description: Invalid Data.
-      summary: Funcion to decommission a server
-  /serverlesspoc:
-    post:
-      requestBody:
-        content:
-          application/json:
-            schema:
-              type: object
-      responses:
-        "200":
-          description: success
-  /testgradle:
-    post:
-      requestBody:
-        content:
-          application/json:
-            schema:
-              properties:
-                serverName:
-                  type: string
-              type: object
-        description: user to add to the system
-      responses:
-        5XX:
-          description: Internal API Error.
-        "200":
-          content:
-            text/plain:
-              schema:
-                example: Server Decommission successful
-                type: string
-          description: Decommission message
-        "404":
-          description: Invalid server.
-        "417":
-          description: Invalid Data.
-      summary: Funcion to decommission a server
-  /testing:
-    post:
-      requestBody:
-        content:
-          application/json:
-            schema:
-              type: object
-      responses:
-        "200":
-          description: success
-  /testopenfaasfunction:
-    post:
-      requestBody:
-        content:
-          application/json:
-            schema:
-              type: object
-      responses:
-        "200":
-          description: success
-servers:
-- url: https://ofgateway.com/function
