--- test-files/files/out-files/test3.yaml
@@ -1,9 +1,0 @@
-info:
-  license:
-    name: MIT
-  title: Openfaas Sandbox
-  version: 1.0.0
-openapi: 3.0.0
-paths: {}
-servers:
-- url: https://ofgateway.com/function
